reduce duplicated test cases
