+++ Tests/SwiftAvroCoreTests/AvroIPCContextAndSessionTest.swift
