+++ Tests/SwiftAvroCoreTests/AvroJSONEncoderTest.swift
@@ -16,6 +16,35 @@ private func jsonString(from avro: Avro, encoding value: some Encodable) throws 
     String(decoding: try avro.encode(value), as: UTF8.self)
 }
 
+@Suite("JSONValue Equality")
+struct JSONValueEqualityTests {
+    @Test("JSONValue equatable works within same type")
+    func equatableSameType() {
+        #expect(JSONValue.null == .null)
+        #expect(JSONValue.bool(true) == .bool(true))
+        #expect(JSONValue.bool(true) != .bool(false))
+        #expect(JSONValue.int(42) == .int(42))
+        #expect(JSONValue.int(42) != .int(43))
+        #expect(JSONValue.double(3.14) == .double(3.14))
+        #expect(JSONValue.double(3.14) != .double(2.71))
+        #expect(JSONValue.string("a") == .string("a"))
+        #expect(JSONValue.string("a") != .string("b"))
+        #expect(JSONValue.array([.int(1)]) == .array([.int(1)]))
+        #expect(JSONValue.array([.int(1)]) != .array([.int(2)]))
+        #expect(JSONValue.object(["a": .int(1)]) == .object(["a": .int(1)]))
+        #expect(JSONValue.object(["a": .int(1)]) != .object(["b": .int(1)]))
+    }
+
+    @Test("JSONValue equatable across different types returns false")
+    func equatableAcrossTypes() {
+        #expect(JSONValue.int(1) != .string("1"))
+        #expect(JSONValue.bool(true) != .int(1))
+        #expect(JSONValue.null != .int(0))
+        #expect(JSONValue.string("") != .null)
+        #expect(JSONValue.array([]) != .object([:]))
+    }
+}
+
 // MARK: - Primitives (SingleValueEncodingContainer)
 
 @Suite("AvroJSONEncoder – SingleValue primitives")
