+++ Tests/SwiftAvroCoreTests/AvroPrimitiveSizerAndEncoderTest.swift
